requirements.txt
flask==2.0.1
waitress==2.0.0
pyjwt==2.3.0
python-dotenv==0.19.2
werkzeug==2.0.1